+++ Fillit/fillit.h
@@ -6,7 +6,7 @@
 /*   By: jkuusist <jkuusist@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/26 10:32:01 by jkuusist          #+#    #+#             */
-/*   Updated: 2019/11/28 13:47:57 by jkuusist         ###   ########.fr       */
+/*   Updated: 2019/11/29 10:16:28 by lharvey          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 # define FILLIT_H
 
 unsigned char	*check_input(int fd);
-unsigned int	*create_block_array(unsigned char *string);
-unsigned int	*binary_checker(unsigned int *semi_valid_array);
-unsigned int	str_to_bin(unsigned char *block);
+unsigned short	*create_block_array(unsigned char *string);
+unsigned short	*binary_checker(unsigned short *semi_valid_array);
+unsigned short	str_to_bin(unsigned char *block);
 
 #endif
